ic_bls12_381
Module hash_to_curve

This module implements hash_to_curve, hash_to_field and related hashing primitives for use with BLS signatures.

Structs

ExpandMsgXmd

Constructor for expand_message_xmd for a given digest hash function, message, DST, and output length.

ExpandMsgXof

A generator for the output of expand_message_xof for a given extendable hash function, message, DST, and output length.

Traits

ExpandMessage

A trait for message expansion methods supported by hash-to-curve.

ExpandMessageState

Trait for types implementing the expand_message interface for hash_to_field.

HashToCurve

Implementation of random oracle maps to the curve.

HashToField

Enables a byte string to be hashed into one or more field elements for a given curve.

InitExpandMessage

Trait for constructing a new message expander.

MapToCurve

Allow conversion from the output of hashed or encoded input into points on the curve