ic_bls12_381::hash_to_curve
Trait HashToCurve

pub trait HashToCurve<X: ExpandMessage>: MapToCurve + for<'a> Add<&'a Self, Output = Self> {
    // Provided methods
    fn hash_to_curve(message: impl AsRef<[u8]>, dst: &[u8]) -> Self { ... }
    fn encode_to_curve(message: impl AsRef<[u8]>, dst: &[u8]) -> Self { ... }
}

Implementation of random oracle maps to the curve.

Provided Methods

fn hash_to_curve(message: impl AsRef<[u8]>, dst: &[u8]) -> Self

Implements a uniform encoding from byte strings to elements of Self.

This function is suitable for most applications requiring a random oracle returning points in Self.

fn encode_to_curve(message: impl AsRef<[u8]>, dst: &[u8]) -> Self

Implements a non-uniform encoding from byte strings to elements of Self.

The distribution of its output is not uniformly random in Self: the set of possible outputs of this function is only a fraction of the points in Self, and some elements of this set are more likely to be output than others. See section 10.1 of draft-irtf-cfrg-hash-to-curve-12 for a more precise definition of encode_to_curve’s output distribution.

Object Safety

This trait is not object safe.