ic_bls12_381::hash_to_curve
Trait MapToCurve

pub trait MapToCurve: Sized {
    type Field: Copy + Default + HashToField;

    // Required methods
    fn map_to_curve(elt: &Self::Field) -> Self;
    fn clear_h(&self) -> Self;
}

Allow conversion from the output of hashed or encoded input into points on the curve

Required Associated Types

type Field: Copy + Default + HashToField

The field element type.

Required Methods

fn map_to_curve(elt: &Self::Field) -> Self

Maps an element of the finite field Self::Field to a point on the curve Self.

fn clear_h(&self) -> Self

Clears the cofactor, sending a point on curve E to the target group (G1/G2).

Object Safety

This trait is not object safe.