ic_bls12_381
Struct G1Projective

pub struct G1Projective { /* private fields */ }

This is an element of $\mathbb{G}_1$ represented in the projective coordinate space.

Implementations

impl G1Projective

pub fn identity() -> G1Projective

Returns the identity of the group: the point at infinity.

pub fn generator() -> G1Projective

Returns a fixed generator of the group. See notes::design for how this generator is chosen.

pub fn double(&self) -> G1Projective

Computes the doubling of this point.

pub fn add(&self, rhs: &G1Projective) -> G1Projective

Adds this point to another point.

pub fn add_mixed(&self, rhs: &G1Affine) -> G1Projective

Adds this point to another point in the affine model.

pub fn clear_cofactor(&self) -> G1Projective

Multiplies by $(1 - z)$, where $z$ is the parameter of BLS12-381, which suffices to clear the cofactor and map elliptic curve points to elements of $\mathbb{G}_1$.

pub fn batch_normalize(p: &[Self], q: &mut [G1Affine])

Converts a batch of G1Projective elements into G1Affine elements. This function will panic if p.len() != q.len().

pub fn is_identity(&self) -> Choice

Returns true if this element is the identity (the point at infinity).

pub fn is_on_curve(&self) -> Choice

Returns true if this point is on the curve. This should always return true unless an “unchecked” API was used.