ic_bls12_381
Struct Scalar

pub struct Scalar(/* private fields */);

Represents an element of the scalar field $\mathbb{F}_q$ of the BLS12-381 elliptic curve construction.

Implementations

impl Scalar

pub const fn zero() -> Scalar

Returns zero, the additive identity.

pub const fn one() -> Scalar

Returns one, the multiplicative identity.

pub const fn double(&self) -> Scalar

Doubles this field element.

pub fn from_bytes(bytes: &[u8; 32]) -> CtOption<Scalar>

Attempts to convert a little-endian byte representation of a scalar into a Scalar, failing if the input is not canonical.

pub fn to_bytes(&self) -> [u8; 32]

Converts an element of Scalar into a byte representation in little-endian byte order.

pub fn from_bytes_wide(bytes: &[u8; 64]) -> Scalar

Converts a 512-bit little endian integer into a Scalar by reducing by the modulus.

pub const fn from_raw(val: [u64; 4]) -> Self

Converts from an integer represented in little endian into its (congruent) Scalar representation.

pub const fn square(&self) -> Scalar

Squares this element.

pub fn pow(&self, by: &[u64; 4]) -> Self

Exponentiates self by by, where by is a little-endian order integer exponent.

pub fn pow_vartime(&self, by: &[u64; 4]) -> Self

Exponentiates self by by, where by is a little-endian order integer exponent.

This operation is variable time with respect to the exponent. If the exponent is fixed, this operation is effectively constant time.

pub fn invert(&self) -> CtOption<Self>

Computes the multiplicative inverse of this element, failing if the element is zero.

pub const fn mul(&self, rhs: &Self) -> Self

Multiplies rhs by self, returning the result.

pub const fn sub(&self, rhs: &Self) -> Self

Subtracts rhs from self, returning the result.

pub const fn add(&self, rhs: &Self) -> Self

Adds rhs to self, returning the result.

pub const fn neg(&self) -> Self

Negates self.